ggstackplot: Create Overlapping Stacked Plots

Easily create overlapping grammar of graphics plots for scientific data visualization. This style of plotting is particularly common in climatology and oceanography research communities.

Details

Have you ever wanted to create (partly) overlapping line plots with matched color-coding of the data and axes? These kinds of plots are common in climatology and oceanography research but there is not an easy way to create them with ggplot facets. The ggstackplot package builds on ggplot2 to provide a straightforward approach to building these kinds of plots while retaining the powerful grammar of graphics approach of ggplots. Check out the functionality provided by ggstackplots at https://ggstackplot.kopflab.org

See also

Author

Maintainer: Sebastian Kopf sebastian.kopf@colorado.edu (ORCID) [copyright holder]

Authors:

Tristan Caro tristan.caro@colorado.edu (ORCID)

Jamie McFarlin jamie.mcfarlin@uwyo.edu (ORCID)

Jon Raberg Jonathan.Raberg@colorado.edu (ORCID)